+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
 ## <next>
+* Added a simple `Primitive.Spinner`
+* Added a header prop to `Content.Card`
 
 ## 2.4.0
 * Added a loading spinner to `Content.Card`